server is nonblocking and equip_id is broadcasted upon player entry
