+++ main.py
@@ -1,11 +1,12 @@
 from game import Game
 from database import Database
+from server import Server
 import sys
 
 # Create database object
 db = Database()
 # Create server object
-server = None
+server = Server()
 # Create game object
 game = Game(db, server)
